style: all moved to .h
