+++ DragonBurn-kernel/DragonBurn-kernel.vcxproj
@@ -99,16 +99,17 @@
     <ClCompile Include="utils.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="include\cfg.hpp" />
-    <ClInclude Include="include\intel_driver.hpp" />
-    <ClInclude Include="include\intel_driver_resource.hpp" />
-    <ClInclude Include="include\kdmapper.hpp" />
-    <ClInclude Include="include\logger.hpp" />
-    <ClInclude Include="include\nt.hpp" />
-    <ClInclude Include="include\portable_executable.hpp" />
-    <ClInclude Include="include\service.hpp" />
-    <ClInclude Include="include\utils.hpp" />
-    <ClInclude Include="include\web_api.hpp" />
+    <ClInclude Include="include\cfg.h" />
+    <ClInclude Include="include\crypto.h" />
+    <ClInclude Include="include\intel_driver.h" />
+    <ClInclude Include="include\intel_driver_resource.h" />
+    <ClInclude Include="include\kdmapper.h" />
+    <ClInclude Include="include\logger.h" />
+    <ClInclude Include="include\nt.h" />
+    <ClInclude Include="include\portable_executable.h" />
+    <ClInclude Include="include\service.h" />
+    <ClInclude Include="include\utils.h" />
+    <ClInclude Include="include\web_api.h" />
     <ClInclude Include="resource.h" />
   </ItemGroup>
   <ItemGroup>
