+++ DragonBurn-kernel/include/intel_driver_resource.h
