+++ DragonBurn-kernel/portable_executable.cpp
@@ -1,4 +1,4 @@
-#include "portable_executable.hpp"
+#include "portable_executable.h"
 
 #include <Windows.h>
 #include <string>
